fix tests and gitignore

+++ src/main/java/hexlet/code/controller/UserController.java
@@ -6,12 +6,17 @@ import hexlet.code.dto.UserUpdateDTO;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.JsonNullableMapper;
 import hexlet.code.mapper.UserMapper;
+import hexlet.code.model.User;
 import hexlet.code.repository.UserRepository;
 import hexlet.code.utils.PasswordHashing;
+import hexlet.code.utils.UserUtils;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -36,6 +41,9 @@ public class UserController {
 
     @Autowired
     private UserMapper userMapper;
+
+    /*@Autowired
+    private UserUtils userUtils;*/
     @GetMapping("")
     public ResponseEntity<List<UserDTO>> index() {
         var users = userRepository.findAll();
@@ -58,8 +66,25 @@ public class UserController {
 
     @DeleteMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
+    @PreAuthorize("@userUtils.isUserAllow(#id)")
     public void delete(@PathVariable Long id) {
+/*        User UserEnter =  userUtils.getCurrentUser();
+        var UniqeName = UserEnter.getUsername();*/
+        //var userUtils = new UserUtils();
+        //var userConnect = userUtils.getCurrentUser();
+/*        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null || !authentication.isAuthenticated()) {
+            return;
+        }
+        var email = authentication.getName();*/
+/*        var userUtils = new UserUtils();
+        var userConnect = userUtils.getCurrentUser();*/
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        var email = authentication.getName();
         var user = userRepository.findById(id).get();
+        if (!email.equals(user.getUsername())) {
+            return;
+        }
         if (user.getTasks().isEmpty()) {
             userRepository.deleteById(id);
         }
@@ -81,6 +106,7 @@ public class UserController {
 
     @PutMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.OK)
+    @PreAuthorize("@userUtils.isUserAllow(#id)")
     public UserDTO update(@PathVariable Long id, @Valid @RequestBody UserUpdateDTO productData) {
         var user =  userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
